Fix stockLevel type: int instead of float (API expects Java Long)

+++ target_optiply/unified_schemas/products.py
@@ -11,7 +11,7 @@ class ProductSchema(OptiplyBaseSchema):
 
     # Mandatory
     name: str
-    stockLevel: float
+    stockLevel: int
     unlimitedStock: bool = False
 
     # Optional
